+++ Nos/Views/Home/HomeFeedView.swift
@@ -11,8 +11,6 @@ struct HomeFeedView: View {
     @Environment(CurrentUser.self) var currentUser
     @ObservationIgnored @Dependency(\.analytics) private var analytics
 
-    @FetchRequest private var authors: FetchedResults<Author>
-    
     @State private var lastRefreshDate = Date(
         timeIntervalSince1970: Date.now.timeIntervalSince1970 + Double(Self.staticLoadTime)
     )
@@ -30,24 +28,11 @@ struct HomeFeedView: View {
 
     let user: Author
 
-    @State private var stories: [Author] = []
-    @State private var selectedStoryAuthor: Author?
-    @State private var storiesCutoffDate = Calendar.current.date(byAdding: .day, value: -2, to: .now)!
-    
     @State private var showRelayPicker = false
     @State private var selectedRelay: Relay? 
 
-    private var isShowingStories: Bool {
-        selectedStoryAuthor != nil
-    }
-    
     init(user: Author) {
         self.user = user
-        _authors = FetchRequest(
-            fetchRequest: user.followedWithNewNotes(
-                since: Calendar.current.date(byAdding: .day, value: -2, to: .now)!
-            )
-        )
     }
     
     var homeFeedFetchRequest: NSFetchRequest<Event> {
@@ -65,30 +50,11 @@ struct HomeFeedView: View {
     var navigationBarTitle: LocalizedStringResource {
         if let relayName = selectedRelay?.host {
             LocalizedStringResource(stringLiteral: relayName)
-        } else if isShowingStories {
-            .localizable.stories
         } else {
             .localizable.accountsIFollow
         }
     }
-    
-    /// Downloads the data we need to show stories. 
-    func downloadStories() async {
-        relaySubscriptions.removeAll()
-        
-        let followedKeys = await Array(currentUser.socialGraph.followedKeys)
-        
-        if !followedKeys.isEmpty {
-            let textFilter = Filter(
-                authorKeys: followedKeys, 
-                kinds: [.text, .delete, .repost, .longFormContent, .report], 
-                since: storiesCutoffDate
-            )
-            let textSubs = await relayService.fetchEvents(matching: textFilter)
-            relaySubscriptions.append(textSubs)
-        }
-    }
-    
+
     var body: some View {
         ZStack {
             PagedNoteListView(
@@ -98,16 +64,7 @@ struct HomeFeedView: View {
                 context: viewContext,
                 tab: .home,
                 header: {
-                    Group {
-                        if selectedRelay == nil {
-                            AuthorStoryCarousel(
-                                authors: $stories, 
-                                selectedStoryAuthor: $selectedStoryAuthor
-                            )
-                        } else {
-                            EmptyView()
-                        }
-                    }
+                    EmptyView()
                 },
                 emptyPlaceholder: { _ in
                     VStack {
@@ -118,24 +75,10 @@ struct HomeFeedView: View {
                 },
                 onRefresh: {
                     lastRefreshDate = .now
-                    storiesCutoffDate = Calendar.current.date(byAdding: .day, value: -2, to: lastRefreshDate)!
-                    authors.nsPredicate = user.followedWithNewNotesPredicate(
-                        since: storiesCutoffDate
-                    )
-                    Task { await downloadStories() }
                     return Event.homeFeed(for: user, before: lastRefreshDate)
                 }
             )
             .padding(0)
-            
-            StoriesView(
-                cutoffDate: $storiesCutoffDate,
-                authors: stories,
-                selectedAuthor: $selectedStoryAuthor
-            )
-            .scaleEffect(isShowingStories ? 1 : 0.5)
-            .opacity(isShowingStories ? 1 : 0)
-            .animation(.default, value: selectedStoryAuthor)
             
             if showTimedLoadingIndicator {
                 FullscreenProgressView(
@@ -163,15 +106,11 @@ struct HomeFeedView: View {
             }
         }
         .doubleTapToPop(tab: .home) { _ in
-            if isShowingStories {
-                selectedStoryAuthor = nil
-            } else {
-                NotificationCenter.default.post(
-                    name: .scrollToTop,
-                    object: nil,
-                    userInfo: ["tab": AppDestination.home]
-                )
-            }
+            NotificationCenter.default.post(
+                name: .scrollToTop,
+                object: nil,
+                userInfo: ["tab": AppDestination.home]
+            )
         }
         .background(Color.appBg)
         .toolbar {
@@ -179,46 +118,23 @@ struct HomeFeedView: View {
                 SideMenuButton()
             }
             ToolbarItem(placement: .navigationBarTrailing) {
-                if isShowingStories {
-                    Button {
-                        selectedStoryAuthor = nil
-                    } label: {
-                        Image.stories.rotationEffect(selectedStoryAuthor == nil ? Angle.zero : Angle(degrees: 90))
-                            .animation(.default, value: selectedStoryAuthor)
+                Button {
+                    withAnimation {
+                        showRelayPicker.toggle()
                     }
-                } else {
-                    Button { 
-                        withAnimation {
-                            showRelayPicker.toggle()
-                        }
-                    } label: { 
-                        Image(systemName: "line.3.horizontal.decrease.circle")
-                            .foregroundStyle(Color.secondaryTxt)
-                            .accessibilityLabel(Text(.localizable.filter))
-                    }
-                    .frame(minWidth: 40, minHeight: 40)
+                } label: {
+                    Image(systemName: "line.3.horizontal.decrease.circle")
+                        .foregroundStyle(Color.secondaryTxt)
+                        .accessibilityLabel(Text(.localizable.filter))
                 }
+                .frame(minWidth: 40, minHeight: 40)
             }
         }
         .padding(.top, 1)
         .nosNavigationBar(title: navigationBarTitle)
-        .task {
-            await downloadStories()
-        }
         .onAppear {
             if router.selectedTab == .home {
                 isVisible = true 
-            }
-            if !isShowingStories {
-                stories = authors.map { $0 }
-            }
-        }
-        .onChange(of: isShowingStories) { _, newValue in
-            if newValue {
-                analytics.enteredStories()
-            } else {
-                analytics.closedStories()
-                stories = authors.map { $0 }
             }
         }
         .onDisappear { isVisible = false }